extracted functions to own file
